requirements.txt
matplotlib==3.9.0
numpy==1.26.4
Pillow==10.3.0
scikit_learn==1.5.0
scipy==1.13.1
scikit_image==0.23.2
